pyproject.toml
[project]
name = "geonext-mcp"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "fastmcp>=2.11.0",
  "geopy>=2.4.1",
  "python-dotenv>=1.0",
  "uvicorn[standard]>=0.29",
]

[tool.setuptools.packages.find]
include = ["geonext_mcp*"]   # what to keep
exclude = ["systemd*"]       # what to skip
